refactor: тесты контроллеров разделены на рахные классы, добавлены новые тесты
